+++ common/src/database/schema/sqlite/mod.rs
@@ -10,7 +10,10 @@ use self::{
     _005_add_uri_field_in_subscriptions_table::AddUriFieldInSubscriptionsTable,
     _006_add_content_format_field_in_subscriptions_table::AddContentFormatFieldInSubscriptionsTable,
     _007_add_ignore_channel_error_field_in_subscriptions_table::AddIgnoreChannelErrorFieldInSubscriptionsTable,
-    _008_add_princs_filter_fields_in_subscriptions_table::AddPrincsFilterFieldsInSubscriptionsTable, _009_alter_outputs_format::AlterOutputsFormat, _010_add_revision_field_in_subscriptions_table::AddRevisionFieldInSubscriptionsTable,
+    _008_add_princs_filter_fields_in_subscriptions_table::AddPrincsFilterFieldsInSubscriptionsTable,
+    _009_alter_outputs_format::AlterOutputsFormat,
+    _010_add_revision_field_in_subscriptions_table::AddRevisionFieldInSubscriptionsTable,
+    _011_add_locale_fields_in_subscriptions_table::AddLocaleFieldsInSubscriptionsTable,
 };
 
 mod _001_create_subscriptions_table;
@@ -23,6 +26,7 @@ mod _007_add_ignore_channel_error_field_in_subscriptions_table;
 mod _008_add_princs_filter_fields_in_subscriptions_table;
 mod _009_alter_outputs_format;
 mod _010_add_revision_field_in_subscriptions_table;
+mod _011_add_locale_fields_in_subscriptions_table;
 
 pub fn register_migrations(sqlite_db: &mut SQLiteDatabase) {
     sqlite_db.register_migration(Arc::new(CreateSubscriptionsTable));
@@ -35,4 +39,5 @@ pub fn register_migrations(sqlite_db: &mut SQLiteDatabase) {
     sqlite_db.register_migration(Arc::new(AddPrincsFilterFieldsInSubscriptionsTable));
     sqlite_db.register_migration(Arc::new(AlterOutputsFormat));
     sqlite_db.register_migration(Arc::new(AddRevisionFieldInSubscriptionsTable));
+    sqlite_db.register_migration(Arc::new(AddLocaleFieldsInSubscriptionsTable));
 }